+++ src/Laravel/configs/forge.php
@@ -1,0 +1,25 @@
+<?php
+
+return [
+    /**
+     * Your Forge API token.
+     */
+    'token' => env('FORGE_TOKEN'),
+
+    /**
+     * Here you may specify default credentials for specific providers.
+     *
+     * Provider shortcodes:
+     *
+     * - DigitalOcean: "ocean2"
+     * - Linode: "linode"
+     * - AWS: "aws"
+     */
+    'default_credentials' => [
+
+        // 'ocean2' => 12345,
+        // 'linode' => 54321,
+        // 'aws' => 11111,
+
+    ],
+];